requirements.txt
fastapi==0.117.1
uvicorn==0.32.1
gunicorn==23.0.0
python-multipart==0.0.12
xmltodict==0.13.0
python-dotenv==1.0.1
supabase==2.19.0
google-generativeai==0.8.3
google-genai==1.38.0
PyPDF2==3.0.1
pdfplumber==0.11.4